* Memory fixes. * Bug fixes. * Realtime Dock icon progressbar. * Added NSNumber methods to become a better 64-bit citizen.
git-svn-id: https://burn-osx.svn.sourceforge.net/svnroot/burn-osx/trunk@518 df5e183e-a46d-47f3-acf2-86dc86df8d71

+++ burn/Source/KWConverter.h
@@ -25,9 +25,9 @@
 	//Input file values
 	NSInteger inputWidth;
 	NSInteger inputHeight;
-	float inputFps;
+	CGFloat inputFps;
 	NSInteger inputTotalTime;
-	float inputAspect;
+	CGFloat inputAspect;
 	//inputFormat: 0 = normal; 1 = dv; 2 = mpeg2
 	NSInteger inputFormat;
 
@@ -84,7 +84,7 @@
 
 //Other actions
 - (NSInteger)convertToEven:(NSString *)number;
-- (NSInteger)getPadSize:(float)size withAspect:(NSSize)aspect withTopBars:(BOOL)topBars;
+- (NSInteger)getPadSize:(CGFloat)size withAspect:(NSSize)aspect withTopBars:(BOOL)topBars;
 - (BOOL)remuxMPEG2File:(NSString *)path outPath:(NSString *)outFile;
 - (BOOL)canCombineStreams:(NSString *)path;
 - (BOOL)combineStreams:(NSString *)path atOutputPath:(NSString *)outputPath;
